Do not start launcher and panel at "Panel" stage
This is a workaround because neither launcher nor panel registers with the SM.
gnome-session requires applications started at Panel stage to register and will
wait with a timeout if they don't, while it does not require such registration
when the app is started at the "Application" stage (the default one).

We probably want to register in the long run, as I think it will speed up
logout, but note that even if it registers, gnome-panel has been patched to
start at the "Application" stage, so we are not too weird.