GnomeBackground: fixed 'wallpaper' and 'scaled' modes (cf. launchpad bug #689531).
'wallpaper' was broken because the sourceSize was set and taken into account even
if the image's size was smaller (http://bugreports.qt.nokia.com/browse/QTBUG-16860).
Thanks to the newly introduced CacheEffect, there is no need to prescale the Image
anymore therefore using sourceSize can be avoided. Unfortunately BlendedImageProvider
was failing if sourceSize was unset. This is now fixed.

'scaled' was broken because the background color was invisible whenever the image
was visible. This optimization is no longer necessary thanks to the CacheEffect.