[dash] Fixed synchronisation of GridView in places/RendererGrid.qml with its parent flickable.
The conditions on flickable.height and totalHeight were in the wrong place: we always
want to set 'y' and 'contentY'. Only in some cases they should be set to 0.
Handle extra cases where synchronisation is required: when the model changes and
when the number of items in the model changes.