Fix the "application://" URLs passed by the backend.
The backend builds URLs of the form "application://filename.desktop". Strictly speaking, this is not a valid URL as the filename takes the position of the host.
This is problematic because the QUrl constructor converts the host name to lower case to conform to the Nameprep RFC (see http://doc.qt.nokia.com/qurl.html#FormattingOption-enum), thus loosing the case of the filename.