Moving to the new UnityCore interface.
An IndicatorEntry now has a priority property which we must use to
order the indicators in the panel. Basically we must use just two
Horizontal layouts to box all the indicators. One is used for menu items
the other for all the indicators, which can be ordered independently
from their origin.