[panel] Workaround QConf incomplete destruction
Ensure that the QConf object stays alive as long as the PanelManager is. This
is to workaround an issue with QConf's destructor, which doesn't disconnect
from the DConf's "notify" signal that can later be emitted while running the
event loop, and would therefore attempt to operate on an already destroyed
object.